+++ src/main/app.ts
@@ -5,6 +5,7 @@
  */
 
 import { app, BrowserWindow, Tray, Menu, globalShortcut, nativeImage, ipcMain, session as electronSession, powerSaveBlocker, shell } from 'electron';
+import * as os from 'os';
 import * as path from 'path';
 import * as fs from 'fs';